+++ src/smtp/send.rs
@@ -32,6 +32,11 @@ impl Smtp {
         message: &[u8],
         rowid: i64,
     ) -> Result<()> {
+        // Notify ratelimiter about sent message regardless of whether quota is exceeded or not.
+        // Checking whether sending is allowed for low-priority messages should be done by the
+        // caller.
+        context.ratelimit.write().await.send();
+
         let message_len_bytes = message.len();
 
         let mut chunk_size = DEFAULT_MAX_SMTP_RCPT_TO;